second commit

+++ controllers/sauce.js
@@ -1,6 +1,5 @@
 const Sauce = require("../models/Sauce");
 const fs = require("fs");
-const { use } = require("../app");
 
 exports.createSauce = (req, res, next) => {
       const sauceObject = JSON.parse(req.body.sauce);
@@ -29,16 +28,19 @@ exports.createSauce = (req, res, next) => {
 };
 
 exports.modifySauce = (req, res, next) => {
+      // console.log(req.params.id);
       Sauce.findOne({ _id: req.params.id })
-
             .then((sauce) => {
                   const filename = sauce.imageUrl.split("/images/")[1];
-                  console.log("reqparamsid:", req.params.id);
-                  console.log("sauce id:", sauce._id);
-                  console.log("sauce userid:", sauce.userId);
+                  // console.log("reqparamsid:", req.params.id);
+                  // console.log("sauce id:", sauce._id);
+                  // console.log("sauce userid:", sauce.userId);
+                  // console.log("normal userid:", req.file.userId);
                   if (req.params.id == sauce._id) {
+                        // check
+
                         console.log("filename:", filename);
-                        console.log("req.file", req.file.filename);
+
                         if (filename != req.file.filename) {
                               fs.unlink(`images/${filename}`, () => {
                                     const sauceObject = req.file
@@ -68,29 +70,38 @@ exports.modifySauce = (req, res, next) => {
                                           );
                               });
                         }
+                  } else {
+                        res.status(403).json({ error: "unauthorized request" });
                   }
             })
-
-            .catch((error) => res.status(400).json({ error }));
+            .catch((error) => res.status(500).json({ error })); // check
 };
 
 exports.deleteSauce = (req, res, next) => {
       Sauce.findOne({ _id: req.params.id })
-            .then((sauce) => {
-                  const filename = sauce.imageUrl.split("/images/")[1];
 
-                  fs.unlink(`images/${filename}`, () => {
-                        Sauce.deleteOne({ _id: req.params.id })
-                              .then(() =>
-                                    res
-                                          .status(200)
-                                          .json({ message: "Sauce supprimée" })
-                              )
-                              .catch((error) =>
-                                    res.status(400).json({ error })
-                              );
-                  });
+            .then((sauce) => {
+                  if (req.params.id == sauce._id) {
+                        const filename = sauce.imageUrl.split("/images/")[1];
+
+                        fs.unlink(`images/${filename}`, () => {
+                              Sauce.deleteOne({ _id: req.params.id })
+                                    .then(() =>
+                                          res
+                                                .status(200)
+                                                .json({
+                                                      message: "Sauce supprimée",
+                                                })
+                                    )
+                                    .catch((error) =>
+                                          res.status(400).json({ error })
+                                    );
+                        });
+                  } else {
+                        res.status(403).json({ error: "unauthorized request" });
+                  }
             })
+
             .catch((error) => res.status(500).json({ error }));
 };
 
@@ -107,78 +118,92 @@ exports.getOneSauce = (req, res, next) => {
 };
 
 exports.likeSauce = (req, res, next) => {
-      if (req.body.like === 1) {
-            // Check
-            console.log(req.body.userId);
-            console.log(req.body.like);
+      Sauce.findOne({ _id: req.params.id }).then((sauce) => {
+            if (
+                  req.body.like === 1 &&
+                  !sauce.usersLiked.includes(req.body.userId)
+            ) {
+                  // Check
+                  console.log("REQ BODY USER ID:", req.body.userId);
+                  console.log("REQ BODY:", req.body);
+                  console.log("REQ BODY LIKE:", req.body.like);
 
-            Sauce.updateOne(
-                  { _id: req.params.id },
-                  {
-                        $inc: { likes: req.body.like++ },
-                        $push: { usersLiked: req.body.userId },
-                  }
-            )
-                  .then((sauce) =>
-                        res.status(200).json({ message: "Ajout Like" })
-                  )
-                  .catch((error) => res.status(400).json({ error }));
-      } else if (req.body.like === -1) {
-            // Check
-            Sauce.updateOne(
-                  { _id: req.params.id },
-                  {
-                        $inc: { dislikes: req.body.like++ * -1 },
-                        $push: { usersDisliked: req.body.userId },
-                  }
-            )
-                  .then((sauce) =>
-                        res.status(200).json({ message: "Ajout Dislike" })
-                  )
-                  .catch((error) => res.status(400).json({ error }));
-      } else {
-            Sauce.findOne({ _id: req.params.id })
-                  .then((sauce) => {
-                        if (sauce.usersLiked.includes(req.body.userId)) {
-                              Sauce.updateOne(
-                                    { _id: req.params.id },
-                                    {
-                                          $pull: {
-                                                usersLiked: req.body.userId,
-                                          },
-                                          $inc: { likes: -1 },
-                                    }
-                              )
-                                    .then((sauce) => {
-                                          res.status(200).json({
-                                                message: "Suppression Like",
-                                          });
-                                    })
-                                    .catch((error) =>
-                                          res.status(400).json({ error })
-                                    );
-                        } else if (
-                              sauce.usersDisliked.includes(req.body.userId)
-                        ) {
-                              Sauce.updateOne(
-                                    { _id: req.params.id },
-                                    {
-                                          $pull: {
-                                                usersDisliked: req.body.userId,
-                                          },
-                                          $inc: { dislikes: -1 },
-                                    }
-                              )
-                                    .then((sauce) => {
-                                          res.status(200).json({
-                                                message: "Suppression Dislike",
-                                          });
-                                    })
-                                    .catch((error) =>
-                                          res.status(400).json({ error })
-                                    );
+                  Sauce.updateOne(
+                        { _id: req.params.id },
+                        {
+                              $inc: { likes: req.body.like++ },
+                              $push: { usersLiked: req.body.userId },
                         }
-                  })
-                  .catch((error) => res.status(400).json({ error }));
-      }
+                  )
+
+                        .then((sauce) =>
+                              res.status(200).json({ message: "Ajout Like" })
+                        )
+                        .catch((error) => res.status(400).json({ error }));
+            } else if (
+                  req.body.like === -1 &&
+                  !sauce.usersDisliked.includes(req.body.userId)
+            ) {
+                  // Check
+                  Sauce.updateOne(
+                        { _id: req.params.id },
+                        {
+                              $inc: { dislikes: req.body.like++ * -1 },
+                              $push: { usersDisliked: req.body.userId },
+                        }
+                  )
+                        .then((sauce) =>
+                              res.status(200).json({ message: "Ajout Dislike" })
+                        )
+                        .catch((error) => res.status(400).json({ error }));
+            } else {
+                  Sauce.findOne({ _id: req.params.id })
+                        .then((sauce) => {
+                              if (sauce.usersLiked.includes(req.body.userId)) {
+                                    Sauce.updateOne(
+                                          { _id: req.params.id },
+                                          {
+                                                $pull: {
+                                                      usersLiked:
+                                                            req.body.userId,
+                                                },
+                                                $inc: { likes: -1 },
+                                          }
+                                    )
+                                          .then((sauce) => {
+                                                res.status(200).json({
+                                                      message: "Suppression Like",
+                                                });
+                                          })
+                                          .catch((error) =>
+                                                res.status(400).json({ error })
+                                          );
+                              } else if (
+                                    sauce.usersDisliked.includes(
+                                          req.body.userId
+                                    )
+                              ) {
+                                    Sauce.updateOne(
+                                          { _id: req.params.id },
+                                          {
+                                                $pull: {
+                                                      usersDisliked:
+                                                            req.body.userId,
+                                                },
+                                                $inc: { dislikes: -1 },
+                                          }
+                                    )
+                                          .then((sauce) => {
+                                                res.status(200).json({
+                                                      message: "Suppression Dislike",
+                                                });
+                                          })
+                                          .catch((error) =>
+                                                res.status(400).json({ error })
+                                          );
+                              }
+                        })
+                        .catch((error) => res.status(400).json({ error }));
+            }
+      }); // here
 };
